src/main/java/com/leetcode/P94.java
package com.leetcode;

import java.util.ArrayList;
import java.util.List;

// https://github.com/antop-dev/algorithm/issues/387
public class P94 {

    public List<Integer> inorderTraversal(TreeNode root) {
        List<Integer> result = new ArrayList<>();
        if (root == null) return result;
        inorder(result, root);
        return result;
    }

    private void inorder(List<Integer> result, TreeNode root) {
        if (root == null) return;
        inorder(result, root.left);
        result.add(root.val);
        inorder(result, root.right);
    }

    public static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;

        TreeNode() {
        }

        TreeNode(int val) {
            this.val = val;
        }

        TreeNode(int val, TreeNode left, TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }
    }

}
